requirements.txt
django==1.11.0
djangorestframework==3.6.2
psycopg2==2.7.1
django-rest-swagger==2.2.0
django-storages==1.7.1
gunicorn==19.9.0
requests==2.21.0
simplejson==3.16.0
